Divide main up into several smaller files
